+++ src/lrcalc/lrcoef.rs
@@ -1,5 +1,7 @@
-use super::helper::ivector::{iv_sum, IntVector};
-use super::part::{part_entry, part_length, part_leq, part_valid};
+use super::{
+    ivector::{_IntVector as IntVector, iv_sum},
+    part::{part_entry, part_length, part_leq, part_valid},
+};
 
 #[derive(Clone)]
 struct LRCoefBox {
